refactor: Move OpBuilderBase/TapeBuilder to onnxscript.tape_builder
- Create onnxscript/tape_builder.py as the canonical location for
  OpBuilderBase (ABC) and TapeBuilder (concrete implementation)
- Export OpBuilderBase and TapeBuilder from onnxscript.__init__
- Slim down rewriter/_context.py to re-export and define local aliases
- Optimizer and version converter import directly from tape_builder,
  defining their own local aliases (OptimizerContext, VCContext)
